Add WIKIDP_LANG env for defining the default language.
Initialize the support for another language with an env variable.
Take care of a fallback language when nothing is defined in the default one.
Correct SPARQL queries by putting a space in the join.
Use the wbgetentities API to retrieve labels for properties to get various language
instead of going to the web page with only ENglish in it.
Fill the cache of properties on init.